Make RadioPicker usable with different buttons
